change: change how we define theme color variables
preparation for tailwind migration
